+++ imm-20200930/core/Models/SubtitleStream.cs
@@ -56,7 +56,7 @@ namespace AlibabaCloud.SDK.Imm20200930.Models
         /// </summary>
         [NameInMap("Duration")]
         [Validation(Required=false)]
-        public float? Duration { get; set; }
+        public double? Duration { get; set; }
 
         /// <summary>
         /// Height
@@ -84,7 +84,7 @@ namespace AlibabaCloud.SDK.Imm20200930.Models
         /// </summary>
         [NameInMap("StartTime")]
         [Validation(Required=false)]
-        public float? StartTime { get; set; }
+        public double? StartTime { get; set; }
 
         /// <summary>
         /// Width
